% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_varnames}
\alias{get_varnames}
\title{Extract variable names from terms object, handling specials}
\usage{
get_varnames(Terms)
}
\arguments{
\item{Terms}{a terms object}
}
\description{
Extract variable names from terms object, handling specials
}
\keyword{Internal}

