% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predict.cosinor.lm}
\alias{predict.cosinor.lm}
\title{Predict from a cosinor model}
\usage{
\method{predict}{cosinor.lm}(cosinor.lm, newdata = NULL, type = "mean")
}
\arguments{
\item{cosinor.lm}{An object of class \code{cosinor.lm}}

\item{newdata}{Optional data frame which contains outcome and time values for forming predictions}

\item{type}{Character string specifying type of prediction. Currently the only option is \code{"mean"}.}
}
\description{
Given a time variable and optional covariates, generate predicted values from a cosinor fit. Default
prediction is the mean value, optionally can predict at a given month
}
\examples{
fit <- cosinor.lm(Y ~ time(time) + X + amp.acro(X), data = vitamind)
predict(fit, type = "mean")
}

