% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cosinor.R
\name{print.test_cosinor}
\alias{print.test_cosinor}
\title{Print results of test of cosinor model}
\usage{
\method{print}{test_cosinor}(x, ...)
}
\arguments{
\item{x}{test_cosinor object}

\item{...}{Arguments passed to \code{print}}
}
\description{
Print results of test of cosinor model
}
