% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simulate_cosinor}
\alias{simulate_cosinor}
\title{Simulate data from a cosinor model}
\usage{
simulate_cosinor(n, beta.mean = 2, beta.amp = 0, beta.acro = 0)
}
\arguments{
\item{n}{Sample size}

\item{beta.mean}{Effect on the mean (intercept)}

\item{beta.amp}{Effect on the amplitude}

\item{beta.acro}{Effect on the acrophase}
}
\description{
This function simulates data from a cosinor model with a single covariate,
where the time scale is month, and optionally
allows for single covariate effects on the mean,
amplitude, and acrophase.
}

