% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor.lm.R
\name{update_covnames}
\alias{update_covnames}
\title{Replace covariate names with descriptive text}
\usage{
update_covnames(names)
}
\arguments{
\item{names}{Coefficient names to update}
}
\description{
Replace covariate names with descriptive text
}
