% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vitamind}
\alias{vitamind}
\title{Vitamin D}
\format{
A data frame with 3 variables: \code{X}, \code{Y},
  \code{time}.
}
\usage{
vitamind
}
\description{
Simulated data set to illustrate the cosinor model. \code{Y}
is an outcome variable that varies of time \code{time} according
to a cosine curve. The binary covariate \code{X} is associated with the
mean and amplitude of the cosine curve.
}
\keyword{datasets}
