% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor_analyzer.R
\name{cosinor_analyzer}
\alias{cosinor_analyzer}
\title{Shiny application to demonstrate cosinor fit}
\usage{
cosinor_analyzer(data = NULL)
}
\arguments{
\item{data}{Data frame to analyze}
}
\description{
Given a dataset, specify the outcome, time variable, and optional covariates. The app will then perform a cosinor analysis and plot the results.
}
\examples{
\dontrun{
library(shiny)
cosinor_analyzer(vitamind)
}

}
