% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.cosinor.lm.R
\name{predict.cosinor.lm}
\alias{predict.cosinor.lm}
\title{Predict from a cosinor model}
\usage{
\method{predict}{cosinor.lm}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class \code{cosinor.lm}}

\item{newdata}{Optional new data}

\item{...}{other arguments}
}
\description{
Given a time variable and optional covariates, generate predicted values from
a cosinor fit. Default prediction is the mean value, optionally can predict
at a given month
}
\examples{

fit <- cosinor.lm(Y ~ time(time) + X + amp.acro(X), data = vitamind)
predict(fit)

}
