% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor.lm.R
\name{print.cosinor.lm}
\alias{print.cosinor.lm}
\title{Print cosinor model}
\usage{
\method{print}{cosinor.lm}(x, ...)
}
\arguments{
\item{x}{cosinor.lm object}

\item{...}{passed to summary}
}
\description{
Given an outcome and time variable, fit the cosinor model with optional covariate effects.
}
