% Generated by roxygen2 (4.0.1): do not edit by hand
\name{update_covnames}
\alias{update_covnames}
\title{Replace covariate names with descriptive text}
\usage{
update_covnames(names)
}
\arguments{
\item{names}{Coefficient names to update}
}
\description{
Replace covariate names with descriptive text
}

